IdManagerV_DEFINITIVA/IdManager/requirements.txt
opencv-python
numpy
pillow
pytesseract